Connected to database and make fetch requests on Look for tutor page

+++ Tutoring_Platform/Models/TutorInfo.cs
@@ -15,7 +15,7 @@ namespace Tutoring_Platform.Models
         public int Id { get; set; }
         public int UserId { get; set; }
         public string Status { get; set; } = null!;
-        public double Wage { get; set; }
+        public float Wage { get; set; }
 
         public virtual StudTutorInfo User { get; set; } = null!;
         public virtual ICollection<AppointRequest> AppointRequests { get; set; }
